:sparkles:	feat: 변수 타입 변경

+++ src/main/java/com/example/momentix/domain/events/repository/seats/SeatsRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface SeatsRepository extends
         JpaRepository<Seats, Long>,SeatsRepositoryCustom{
-    Optional<Seats> findBySeatRowAndSeatColAndPlaces_Id(int seatRow, int seatCol, Long id);
+    Optional<Seats> findBySeatRowAndSeatColAndPlaces_Id(Long seatRow, Long seatCol, Long id);
 }
